Fix button functionality and add pill-shaped buttons with sparkles
CRITICAL FIXES:
1. ✅ PILL-SHAPED BUTTONS:
   - Changed border-radius to 'rounded-full' (pill shape)
   - Updated padding to px-6 py-3 (12px 24px)
   - Added scale transform (1.05) when selected
   - Added box-shadow: 0 4px 6px rgba(0,0,0,0.1) on selected buttons
   - Smooth transitions with ease-in-out

2. ✅ BUTTON FUNCTIONALITY:
   - Each date's buttons work independently
   - State properly tracks each date with unique keys
   - Selected state applies correct colors:
     * Yes: bg-emerald-500 (tropical green)
     * Maybe: bg-amber-400 (sunshine yellow)
     * No: bg-rose-400 (coral pink)
   - Unselected: white bg with 2px gray border

3. ✅ FIXED DATE DISPLAY:
   - Changed from "Fri, Mar 6" to "Friday, March 6"
   - Shows full day of week and month names
   - Dates correctly show March 6-22, 2025 with proper days

4. ✨ ADDED SPARKLES:
   - "✨ WELCOME TO ✨" on landing page
   - "✨ Success! ✨" on confirmation page
   - "Hi [Name]! ✨" on schedule page
   - "✨🌴✨" at bottom of cards
   - Added sparkles to subheadings throughout

All buttons now have beautiful pill shape and work correctly!

+++ lib/dates.ts
@@ -20,8 +20,8 @@ export function formatDate(date: Date): string {
 
 export function formatDisplayDate(date: Date): string {
   return date.toLocaleDateString('en-US', {
-    weekday: 'short',
-    month: 'short',
+    weekday: 'long',
+    month: 'long',
     day: 'numeric',
   });
 }
